move to mostly usize usage
Eventually I will replace the x86_64::VirtAddr/PhysAddr with my own structs, and they'll use `usize`. I don't really see a reason not to, it doesn't hurt code gen, and the types are used more pervasively throughout the rust core & standard libraries.

+++ kernel/src/memory/paging/virtual_addressor.rs
@@ -221,7 +221,7 @@ impl VirtualAddressor {
     }
 
     fn modify_mapped_page(&mut self, page: Page) {
-        let total_memory_pages = (global_total() / 0x1000) as u64;
+        let total_memory_pages = global_total() / 0x1000;
         for index in 0..total_memory_pages {
             self.map(&page.offset(index), &Frame::from_index(index));
         }
